+++ RPi2-Ubuntu/install_new_ros_pkg.sh
@@ -1,7 +1,7 @@
 # Install and build ros_catkin_ws
 cd ~/ros_catkin_ws
-rosinstall_generator ros_comm geometry_msgs --rosdistro indigo --deps --wet-only --exclude roslisp --tar > indigo-custom_ros.rosinstall
+rosinstall_generator ros_comm geometry_msgs --rosdistro $ROS_DISTRO --deps --wet-only --exclude roslisp --tar > $ROS_DISTRO-custom_ros.rosinstall
 
 #
-wstool merge -t src indigo-custom_ros.rosinstall
+wstool merge -t src $ROS_DISTRO-custom_ros.rosinstall
 wstool update -t src